% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{cell.densities}
\alias{cell.densities}
\title{Calculate cell-type (weighted) densities/frequencies around each cell.}
\usage{
cell.densities(
  df_raw,
  spot,
  add_markers = FALSE,
  nbhd_def = spdep::dnearneigh,
  nbhd_pars = list(0, 50),
  density = TRUE,
  weighted = FALSE
)
}
\arguments{
\item{spot}{}

\item{add_markers}{}

\item{nbhd_def}{}

\item{nbhd_pars}{}

\item{density}{}

\item{weighted}{}
}
\value{

}
\description{
Calculate cell-type (weighted) densities/frequencies around each cell.
}
