% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{celltype_heatmap}
\alias{celltype_heatmap}
\title{Create a basic cell type heatmap across all neighborhoods (co-occurrence of cell types in neighborhoods)}
\usage{
celltype_heatmap(nbhds.obj)
}
\arguments{
\item{nbhds.obj}{Neighborhoods object, as returned by sg_to_nbhds}
}
\value{

}
\description{
Create a basic cell type heatmap across all neighborhoods (co-occurrence of cell types in neighborhoods)
}
