% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{celltype_pairsplot}
\alias{celltype_pairsplot}
\title{Title}
\usage{
celltype_pairsplot(nbhds.obj, clusters, cell_types = NULL)
}
\arguments{
\item{nbhds.obj}{}

\item{clusters}{}

\item{columns}{}
}
\value{

}
\description{
Title
}
