% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_brms}
\alias{check_brms}
\title{Check a BRMS model using the DHARMa simulated residuals}
\usage{
check_brms(model, integer = TRUE, plot = TRUE, ...)
}
\arguments{
\item{model}{brms model}

\item{integer}{integer response? (TRUE/FALSE)}

\item{plot}{make plot?}

\item{...}{further arguments for DHARMa::plotResiduals}
}
\value{
a DHARMa object
}
\description{
Check a BRMS model using the DHARMa simulated residuals
}
