% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhds.R
\name{cluster_nbhds}
\alias{cluster_nbhds}
\title{Cluster neighborhoods}
\usage{
cluster_nbhds(
  pat = NULL,
  nbhds.obj = NULL,
  ntype = "gabriel",
  scale = "comp",
  par = NULL,
  spat.dist = NULL,
  blusparam = bluster::NNGraphParam()
)
}
\arguments{
\item{pat}{}

\item{ntype}{}

\item{scale}{}

\item{blusparam}{}
}
\value{

}
\description{
Cluster neighborhoods
}
