% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_ppp}
\alias{create_ppp}
\title{Create point pattern with certain cell types}
\usage{
create_ppp(X, Y, cell_types, keep_types = c("Tregs", "CD8+ T cells"))
}
\arguments{
\item{X}{X coordinates of cells}

\item{Y}{Y coordinates of cells}

\item{cell_types}{string of cell type}

\item{keep_types}{types to retain in ppp}
}
\value{
spatstat ppp object
}
\description{
Create point pattern with certain cell types
}
