% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcors.R
\name{densities.pcors}
\alias{densities.pcors}
\title{Title}
\usage{
densities.pcors(
  df_raw,
  spot,
  add_markers = FALSE,
  densities = TRUE,
  nbhd_def = spdep::dnearneigh,
  nbhd_pars = list(0, 50),
  weight_fun = mgwrsar::bisq_C,
  weight_pars = list(200, 0),
  pcor.fun = pcor
)
}
\arguments{
\item{df_raw}{}

\item{spot}{}

\item{add_markers}{}

\item{densities}{}

\item{nbhd_def}{}

\item{nbhd_pars}{}

\item{weight_fun}{}

\item{weight_pars}{}

\item{pcor.fun}{}
}
\value{

}
\description{
Title
}
