% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{.Cell.densities}
\alias{.Cell.densities}
\title{Calculate cell-type (weighted) densities/frequencies around each cell.}
\usage{
.Cell.densities(
  X,
  Y,
  types,
  CellID = 1:length(X),
  type.levels = levels(types),
  markers = NULL,
  nbhd_def = spdep::dnearneigh,
  nbhd_pars = list(0, 50),
  density = TRUE,
  weighted = FALSE,
  weighting_scheme = "raw"
)
}
\arguments{
\item{X}{X coordinates for each cell}

\item{Y}{Y coordinates for each cell}

\item{types}{cell type for each cell}

\item{CellID}{IDs for each cell}

\item{type.levels}{possible cell types}

\item{markers}{cellular biomarkers to include with densities}

\item{nbhd_def}{neighborhood definition, returns list of neighbors for each cell}

\item{nbhd_pars}{parameters for neighborhood definition}

\item{density}{whether to convert cell type frequencies to densities}

\item{weighted}{whether to weight cells in neighborhood by distance from focal cell}

\item{weighting_scheme}{see "style" argument for spdep::nb2listwdist}
}
\value{

}
\description{
Calculate cell-type (weighted) densities/frequencies around each cell.
}
