% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcors.R
\name{.Pcors}
\alias{.Pcors}
\title{Title}
\usage{
.Pcors(
  x,
  y,
  X,
  weight_fun = mgwrsar::bisq_C,
  weight_pars = list(200, 0),
  pcor.fun = pcor,
  method = "kendall"
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{X}{}

\item{weight_fun}{}

\item{weight_pars}{}

\item{pcor.fun}{}

\item{method}{}
}
\value{

}
\description{
Title
}
