% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest plot for coefficients from tidy summary of model}
\usage{
forest_plot(tb, xintercept = 0)
}
\arguments{
\item{tb}{tidied summary of model}

\item{xintercept}{where to draw vertical reference line}
}
\value{
a ggplot object
}
\description{
Forest plot for coefficients from tidy summary of model
}
