% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhds.R
\name{get_disc_nbhd}
\alias{get_disc_nbhd}
\title{Get disc neighborhood around a point}
\usage{
get_disc_nbhd(pat, point, radius = 50)
}
\arguments{
\item{pat}{a marked ppp object from spatstat}

\item{point}{(x,y) point around which to get nbhd}

\item{radius}{radius of disc}
}
\value{
named vector with count of each cell type
}
\description{
Get disc neighborhood around a point
}
