% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhds.R
\name{get_raster_nbhds}
\alias{get_raster_nbhds}
\title{Get all scanned raster neighborhoods in point pattern}
\usage{
get_raster_nbhds(pat, radius = 50)
}
\arguments{
\item{pat}{marked ppp object from spatstat}

\item{radius}{radius of raster neighborhoods}
}
\value{
dataframe of raster neighborhood composition
}
\description{
Get all scanned raster neighborhoods in point pattern
}
