% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kernel.density}
\alias{kernel.density}
\title{Estimate kernel density in multitype point pattern}
\usage{
kernel.density(df_raw, spot, eps, sigma = NULL, ranges = NULL)
}
\arguments{
\item{df_raw}{}

\item{spot}{}

\item{eps}{resolution of kernel density}

\item{sigma}{kernel type to use}

\item{ranges}{X and Y ranges to use for point pattern}
}
\value{

}
\description{
Estimate kernel density in multitype point pattern
}
