% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_spot_nbhds}
\alias{make_spot_nbhds}
\title{Make geometric neighborhoods dataframe}
\usage{
make_spot_nbhds(df, spot.labels, keep_types, radius = 50)
}
\arguments{
\item{df}{cleaned data frame from Schurch et al}

\item{spot.labels}{spots for which to create neighborhoods}

\item{keep_types}{cell types to keep}

\item{radius}{radius of neighborhood}
}
\value{
dataframe of neighborhood composition
}
\description{
Make geometric neighborhoods dataframe
}
