% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_group_split}
\alias{named_group_split}
\title{Split groups into list with names of each group}
\usage{
named_group_split(.tbl, ..., keep = FALSE)
}
\arguments{
\item{.tbl}{tibble to perform group_by}

\item{...}{additional arguments to dplyr::group_by}
}
\value{
list of groups with names
}
\description{
Split groups into list with names of each group
}
