% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{nbhd_celltype_heatmap}
\alias{nbhd_celltype_heatmap}
\title{Heatmap of neighborhood cluster centroids ("regions") and cell type composition}
\usage{
nbhd_celltype_heatmap(centroids)
}
\arguments{
\item{centroids}{Centroids from cluster_nbhds}
}
\value{

}
\description{
Heatmap of neighborhood cluster centroids ("regions") and cell type composition
}
