% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhds.R
\name{nbhdClusterSweep}
\alias{nbhdClusterSweep}
\title{Sweep cluster and neighborhood parameters and compare using ARI}
\usage{
nbhdClusterSweep(
  pats,
  ntypes = list(list(type = "SIG", pars = NULL), list(type = "RNG", pars = NULL),
    list(type = "gabriel", pars = NULL)),
  scales = c("comp", "global.comp", "standardize"),
  clust_params = list(list(BLUSPARAM = bluster::SNNGraphParam(), optParams = list(k =
    c(5L, 10L, 15L, 20L), cluster.fun = c("louvain", "infomap"))), list(BLUSPARAM =
    bluster::SomParam(centers = 1L), optParams = list(centers = 5:10)))
)
}
\arguments{
\item{ntypes}{}

\item{scales}{}

\item{clust_params}{}

\item{pat}{}
}
\value{

}
\description{
Sweep cluster and neighborhood parameters and compare using ARI
}
