% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialOnco-package.R
\docType{package}
\name{nbhdUtils-package}
\alias{nbhdUtils}
\alias{nbhdUtils-package}
\title{nbhdUtils: Spatial Functions For Neighborhood Analysis Of Tumor Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joel Eliason \email{joelne@umich.edu}

}
\keyword{internal}
