% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{pat.density}
\alias{pat.density}
\title{Create kernel density from multitype point pattern}
\usage{
pat.density(pat, eps, sigma, ...)
}
\arguments{
\item{pat}{}

\item{eps}{}

\item{sigma}{}
}
\value{

}
\description{
Create kernel density from multitype point pattern
}
