% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcors.R
\name{pcor}
\alias{pcor}
\title{Partial Correlations
Taken from ppcor R package}
\usage{
pcor(x, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{data matrix}

\item{method}{one of "pearson", "kendall", "spearman"}
}
\value{
list with element "estimate" containing the pcor matrix
}
\description{
Partial Correlations
Taken from ppcor R package
}
