% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcors.R
\name{pcors.table}
\alias{pcors.table}
\title{Title}
\usage{
pcors.table(pcors, CellID = 1:dim(pcors)[3], full = TRUE)
}
\arguments{
\item{pcors}{}

\item{CellID}{}

\item{full}{}
}
\value{

}
\description{
Title
}
