% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcors.R
\name{ppcor_spcor}
\alias{ppcor_spcor}
\title{Semipartial Correlations
Taken from ppcor R package}
\usage{
ppcor_spcor(x, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{data matrix}

\item{method}{one of "pearson", "kendall", "spearman"}
}
\value{
list with element "estimate" containing the spcor matrix
}
\description{
Semipartial Correlations
Taken from ppcor R package
}
