% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pseudospace_plot}
\alias{pseudospace_plot}
\title{Title}
\usage{
pseudospace_plot(
  nbhds.obj,
  clusters,
  against = "Tregs",
  cell_types = c("Tregs", "CD8+ T cells")
)
}
\arguments{
\item{nbhds.obj}{}

\item{clusters}{}

\item{against}{}

\item{cell_types}{}
}
\value{

}
\description{
Title
}
