% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhds.R
\name{range01}
\alias{range01}
\title{Standardize vector to \link{0,1}}
\usage{
range01(x)
}
\arguments{
\item{x}{vector to standardize}
}
\value{
standardized vector
}
\description{
Standardize vector to \link{0,1}
}
