% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_pcors_df}
\alias{read_pcors_df}
\title{Helper to read in pcors df (removes lower triangle)}
\usage{
read_pcors_df(file, NETWORK_PATH, shave = TRUE, split_on = "Spot")
}
\arguments{
\item{file}{}

\item{NETWORK_PATH}{}

\item{shave}{logical, remove lower triangle?}

\item{split_on}{variable to split on}
}
\value{
tibble of pcors
}
\description{
Helper to read in pcors df (removes lower triangle)
}
