% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_pcors_df_to_list}
\alias{read_pcors_df_to_list}
\title{Read a df of pcors (as produced in the sensitivity analysis nb) and convert to a list of pcor matrices}
\usage{
read_pcors_df_to_list(file, NETWORK_PATH, split_on = "Spot")
}
\arguments{
\item{file}{filename to be read from NETWORK_PATH subfolder}

\item{NETWORK_PATH}{path where networks are stored}

\item{split_on}{variable to split the df on}
}
\value{
list of pcor matrices
}
\description{
Read a df of pcors (as produced in the sensitivity analysis nb) and convert to a list of pcor matrices
}
