% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readObj}
\alias{readObj}
\title{Reads object from RESULTS_PATH or other subfolder}
\usage{
readObj(filename, root = NULL)
}
\arguments{
\item{filename}{}

\item{root}{if not NULL, will read from RESULTS_PATH}
}
\value{
R object
}
\description{
Reads object from RESULTS_PATH or other subfolder
}
