% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_model}
\alias{run_model}
\title{Fit or load model}
\usage{
run_model(fit.model, file, ...)
}
\arguments{
\item{fit.model}{model fitting function}

\item{file}{filename}

\item{...}{arguments to pass to fit.model}
}
\value{
fitted model, from disk if already exists, otherwise will fit
}
\description{
Fit or load model
}
