% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveObj}
\alias{saveObj}
\title{Saves object as an RDS file to RESULTS_PATH subfolder}
\usage{
saveObj(obj, filename, overwrite = TRUE)
}
\arguments{
\item{obj}{R object}

\item{filename}{}
}
\value{

}
\description{
Saves object as an RDS file to RESULTS_PATH subfolder
}
