% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhds.R
\name{sg_to_nbhds}
\alias{sg_to_nbhds}
\title{Extract cell neighborhoods from statgraphs object}
\usage{
sg_to_nbhds(sg, pat)
}
\arguments{
\item{sg}{statgraphs object containing the edges for each node}

\item{pat}{spatstat ppp object that sg was created from}
}
\value{
matrix of neighborhood counts of each cell type
}
\description{
Extract cell neighborhoods from statgraphs object
}
\examples{
X = c(1,1,5)
Y = c(2,4,3)
cell_types = c("Tregs","tumor cells","vasculature")
pat = create_ppp(X,Y,cell_types,keep_types=TRUE)
sg = spatgraphs::spatgraph(pat,type="SIG")
nbhds = sg_to_nbhds(sg,pat)
}
