% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcors.R
\name{spatial_contexts}
\alias{spatial_contexts}
\title{Title}
\usage{
spatial_contexts(coords, nbhd, neighbors = 50, threshold = 0.9)
}
\arguments{
\item{coords}{}

\item{nbhd}{}

\item{neighbors}{}

\item{threshold}{}
}
\value{

}
\description{
Title
}
