% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialOnco-package.R
\docType{package}
\name{spatialOnco-package}
\alias{spatialOnco}
\alias{spatialOnco-package}
\title{spatialOnco: Spatial Functions For Analysis Of Tumor Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joel Eliason \email{joelne@umich.edu}

}
\keyword{internal}
