% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcors.R
\name{weighted_cells}
\alias{weighted_cells}
\title{Title}
\usage{
weighted_cells(
  X,
  Y,
  types,
  CellID = 1:length(X),
  type.levels = levels(types),
  markers = NULL,
  nbhd_def = spdep::dnearneigh,
  nbhd_pars = list(0, 50),
  weighted = TRUE
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{types}{}

\item{CellID}{}

\item{type.levels}{}

\item{markers}{}

\item{nbhd_def}{}

\item{nbhd_pars}{}

\item{weighted}{}
}
\value{

}
\description{
Title
}
