% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_brms}
\alias{check_brms}
\title{Title}
\usage{
check_brms(model, integer = TRUE, plot = TRUE, ...)
}
\arguments{
\item{model}{}

\item{integer}{}

\item{plot}{}

\item{...}{}
}
\value{

}
\description{
Title
}
