% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ppp}
\alias{create_ppp}
\title{Create point pattern with certain cell types}
\usage{
create_ppp(X, Y, cell_types, keep_types = "all")
}
\arguments{
\item{X}{X coordinates of cells}

\item{Y}{Y coordinates of cells}

\item{cell_types}{vector of strings of cell type}

\item{keep_types}{string vector of types to retain in ppp, or "all" if should keep all types}
}
\value{
spatstat ppp object
}
\description{
Create point pattern with certain cell types
}
