% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhds.R
\name{scale_nbhds}
\alias{scale_nbhds}
\title{Scale neighborhood counts matrix}
\usage{
scale_nbhds(nbhds, scale = "comp")
}
\arguments{
\item{nbhds}{neighborhoods matrix from sg_to_nbhds}

\item{scale}{string indicating type of scaling, can be any of
"comp","global.comp","binary", or "standardize"}
}
\value{
list with original nbhds, scaled nbhds and type of scaling
}
\description{
Scale neighborhood counts matrix
}
